7271 Main
In SetSidebarVisibility, the sidebar visibility, if only one sidebar for WIN, is true if any document has win->tocVisible true.
The reason is: otherwise, the position of the tab control will change when switching between documents, and this is bad.

Two choice:
  Enable panel but with sidebar visible when any doc's toc is visible.
  Disable panel (so only one panel), then no such problem.

ps: Maybe just do the test for WIN->panel->gWin().

In ShowDocument, fix the bugs for  UpdateToolbarFindText and UpdateFindbo: change win to winNew.